update: updated springboot version and java imports; refactored thymeleaf config.

+++ src/main/java/com/projeto/teste/projetotestes/model/Salary.java
@@ -1,15 +1,16 @@
 package com.projeto.teste.projetotestes.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.ZoneId;
 
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.PrePersist;
-import javax.persistence.Table;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.PrePersist;
+import jakarta.persistence.Table;
 
 import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.format.annotation.DateTimeFormat.ISO;
@@ -42,7 +43,7 @@ public class Salary {
   private LocalDate createdDate;
 
   @Column(nullable = false, precision = 10, scale = 2)
-  private double value;
+  private BigDecimal value;
 
   @PrePersist
   protected void onCreate() {
