+++ src/test/java/com/projeto/teste/model/SalaryTest.java
@@ -3,6 +3,7 @@ package com.projeto.teste.model;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.Month;
 import java.time.ZoneId;
@@ -35,11 +36,11 @@ class SalaryTest {
     s1.setActive(true);
     s1.setCreatedDate(LocalDate.of(2026, Month.JANUARY, 1));
     s1.setId(1);
-    s1.setValue(1632.45);
+    s1.setValue(BigDecimal.valueOf(1632.45));
 
     assertEquals(1, s1.getId());
     assertEquals(true, s1.isActive());
     assertEquals("01/01/2026", s1.getCreatedDate().format(DateTimeFormatter.ofPattern("dd/MM/yyyy")));
-    assertEquals(1632.45, s1.getValue());
+    assertEquals(BigDecimal.valueOf(1632.45), s1.getValue());
   }
 }
